Group timestamps by IP address.
Aha, here's the function I was looking for. groupByKey will give us a
collection of all the time stamps for each IP address. We should be able
to use this (after sorting?) to iterate through and build the sessions.
(At each step, check if the new record is within $TIMEOUT of the old
one).

Prior to this, I had been scratching my head trying to write a reducer
(passed to reduceByKey) that would somehow get the job done without
relying on any execution ordering.